Adding new features to a sample blog
Only the structure in html and css. In this project, I added new features to a sample blog webpage. The original webpage included a header, main content area with articles, and a footer. I added functionality to the navigation bar so that each link opens in a new tab while retaining the purple background color. I also added a dynamic button to allow users to return to the home screen.